[REFACTOR] suggestion module 하나로 구성 (#213)
