+++ app/build.gradle.kts
@@ -86,8 +86,7 @@ dependencies {
     implementation(project(":feature:mission_detail"))
     implementation(project(":feature:manage_mission"))
     implementation(project(":feature:profile"))
-    implementation(project(":feature:suggestion_dashboard"))
-    implementation(project(":feature:suggestion_detail"))
+    implementation(project(":feature:mission_suggestion"))
     implementation(project(":common-ui"))
 
     implementation(libs.hilt)
