+++ app/src/main/java/com/lgtm/android/navigator/LgtmNavigatorImpl.kt
@@ -12,8 +12,8 @@ import com.lgtm.android.manage_mission.dashboard.DashboardActivity
 import com.lgtm.android.manage_mission.ping_pong_junior.PingPongJuniorActivity
 import com.lgtm.android.mission_detail.MissionDetailActivity
 import com.lgtm.android.mission_recommendation.ui.dashboard.SuggestionDashboardActivity
+import com.lgtm.android.mission_recommendation.ui.detail.SuggestionDetailActivity
 import com.lgtm.android.profile.ProfileActivity
-import com.lgtm.android.suggestion_detail.SuggestionDetailActivity
 import javax.inject.Inject
 
 class LgtmNavigatorImpl @Inject constructor(
