+++ feature/mission_suggestion/src/main/AndroidManifest.xml
