--- feature/suggestion_dashboard/build.gradle.kts
@@ -1,34 +1,0 @@
-@Suppress("DSL_SCOPE_VIOLATION") // TODO: Remove once KTIJ-19369 is fixed
-plugins {
-    alias(libs.plugins.dagger.hilt)
-    alias(libs.plugins.android.library)
-    alias(libs.plugins.kotlin.android)
-    kotlin("kapt")
-}
-
-android {
-    namespace = "com.lgtm.android.mission_recommendation"
-
-    buildFeatures {
-        compose = true
-    }
-
-    composeOptions {
-        kotlinCompilerExtensionVersion = libs.versions.kotlin.compiler.get()
-    }
-}
-
-dependencies {
-    implementation(project(":common-ui"))
-
-    implementation(libs.core.ktx)
-    implementation(platform(libs.kotlin.bom))
-    implementation(libs.appcompat)
-    implementation(libs.material)
-    implementation(libs.kotlin.coroutines)
-    implementation(libs.bundles.androidx.ui.foundation)
-    implementation(libs.constraintlayout)
-    implementation(libs.hilt)
-    implementation(libs.bundles.navigation)
-    kapt(libs.hilt.kapt)
-}
