+++ settings.gradle.kts
@@ -27,5 +27,4 @@ include(":feature:mission_detail")
 include(":feature:manage_mission")
 include(":feature:profile")
 include(":swm-logging")
-include(":feature:suggestion_dashboard")
-include(":feature:suggestion_detail")
+include(":feature:mission_suggestion")
